+++ plugins/modules/gns3_nodes_inventory.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 ANSIBLE_METADATA = {
-    "metadata_version": "1.1",
+    "metadata_version": "1.2",
     "status": ["preview"],
     "supported_by": "community",
 }
@@ -15,7 +15,7 @@ description:
     - "Retrieves nodes inventory information from a GNS3 project"
 requirements: [ gns3fy ]
 author:
-    - David Flores (@netpanda)
+    - David Flores (@davidban77)
 options:
     url:
         description:
@@ -27,6 +27,14 @@ options:
             - TCP port to connect to server REST API
         type: int
         default: 3080
+    user:
+        description:
+            - User to connect to GNS3 server
+        type: str
+    password:
+        description:
+            - Password to connect to GNS3 server
+        type: str
     project_name:
         description:
             - Project name
@@ -62,7 +70,7 @@ import traceback
 
 GNS3FY_IMP_ERR = None
 try:
-    import gns3fy
+    from gns3fy import Gns3Connector, Project
 
     HAS_GNS3FY = True
 except ImportError:
@@ -77,6 +85,8 @@ def main():
         argument_spec=dict(
             url=dict(type="str", required=True),
             port=dict(type="int", default=3080),
+            user=dict(type="str", default=None),
+            password=dict(type="str", default=None, no_log=True),
             project_name=dict(type="str", default=None),
             project_id=dict(type="str", default=None),
         ),
@@ -86,26 +96,29 @@ def main():
         module.fail_json(msg=missing_required_lib("gns3fy"), exception=GNS3FY_IMP_ERR)
     result = dict(changed=False, nodes_inventory=None, total_nodes=None)
 
-    server_url = module.params['url']
-    server_port = module.params['port']
-    project_name = module.params['project_name']
-    project_id = module.params['project_id']
+    server_url = module.params["url"]
+    server_port = module.params["port"]
+    server_user = module.params["user"]
+    server_password = module.params["password"]
+    project_name = module.params["project_name"]
+    project_id = module.params["project_id"]
 
     # Create server session
-    server = gns3fy.Gns3Connector(url=f"{server_url}:{server_port}")
+    server = Gns3Connector(
+        url=f"{server_url}:{server_port}", user=server_user, cred=server_password
+    )
     # Define the project
     if project_name is not None:
-        project = gns3fy.Project(name=project_name, connector=server)
+        project = Project(name=project_name, connector=server)
     elif project_id is not None:
-        project = gns3fy.Project(project_id=project_id, connector=server)
+        project = Project(project_id=project_id, connector=server)
 
     # Retrieve project info
     project.get()
 
     nodes_inventory = project.nodes_inventory()
     result.update(
-        nodes_inventory=nodes_inventory,
-        total_nodes=len(nodes_inventory.keys())
+        nodes_inventory=nodes_inventory, total_nodes=len(nodes_inventory.keys())
     )
 
     module.exit_json(**result)